analysis.m is the script to create an excel file stored in the folder preProcessing\blind\cleanedData. It saves the data in 2 sheets. One is for collecting all the data trial per trial, the ohter for collecting average values of data.
centersVSresolution.m is useful for knowing the position of centers depending on the kind of resolution:1 (matrix 2x2), 2(2x3) and 3(3x3).

data_analysis.R is for analyzing data about the arrows pre and post test.

extractActiveTrial.m is for extracting only the portion of trials in which the timer was active

extractBehaviorTrial.m is for calculate some parameters from the files of exploration 

statistical_analyses.R is for analyzing data coming from the excel created with analysis.m script